Fix: New Project Button Logic
