+++ Clients/src/presentation/pages/Home/1.0Home/index.tsx
@@ -33,6 +33,7 @@ import { getAuthToken } from "../../../../application/redux/auth/getAuthToken";
 import { GetMyOrganization } from "../../../../application/repository/organization.repository";
 import { getTierFeatures } from "../../../../application/repository/tiers.repository";
 import { Tier } from "../../../../domain/types/Tiers";
+import { useSubscriptionData } from "../../../../application/hooks/useSubscriptionData";
 
 
 const Home = () => {
@@ -53,8 +54,6 @@ const Home = () => {
 
   const [projects, setProjects] = useState<Project[]>([]);
   const { dashboard, fetchDashboard } = useDashboard();
-  const [organizationTierId, setOrganizationTierId] = useState<number | null>(null);
-  const [tierFeatures, setTierFeatures] = useState<Tier | null>(null);
 
 
   useEffect(() => {
@@ -96,40 +95,16 @@ const Home = () => {
     setRefreshProjectsFlag((prev) => !prev);
   };
 
-  const userToken = extractUserToken(getAuthToken());
-  const organizationId = userToken?.organizationId;
-
-  useEffect(() => {
-    const fetchOrganizationTierId = async () => {
-      const organization = await GetMyOrganization({
-        routeUrl: `/organizations/${organizationId}`,
-      });
-      const org = organization.data.data;
-      setOrganizationTierId(org.subscription_id);
-    }
-
-    fetchOrganizationTierId();
-  }, [organizationId]);
-
-  useEffect(() => {
-    const fetchTierFeatures = async () => {
-      const features = await getTierFeatures({
-        tierId: organizationTierId || 1,
-        routeUrl: "/tiers",
-      });
-      setTierFeatures(features.data);
-    }
-    fetchTierFeatures();
-  }, [organizationTierId]);
+  const { tierFeatures } = useSubscriptionData();
 
   const isDisabledLogic = () => {
-    if (dashboard?.projects && tierFeatures?.projects) {
+    if (dashboard?.projects && tierFeatures?.data.projects) {
       // If tierFeatures.projects is 0, it means unlimited projects
-      if (tierFeatures.projects === 0) {
+      if (tierFeatures.data.projects === 0) {
         return !allowedRoles.projects.create.includes(userRoleName);
       }
       // Otherwise, check if current projects count has reached the limit
-      return dashboard.projects >= tierFeatures.projects || !allowedRoles.projects.create.includes(userRoleName);
+      return dashboard.projects >= tierFeatures.data.projects || !allowedRoles.projects.create.includes(userRoleName);
     }
     return !allowedRoles.projects.create.includes(userRoleName);
   }
